[#32] Post Install script and download WordPress
