+++ bin/composer-scripts/ProjectEvents/PostInstallScript.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * Perform some post-install actions with Composer.
+ */
+
+namespace Viget\ComposerScripts\ProjectEvents;
+
+use Composer\Script\Event;
+use Viget\ComposerScripts\ComposerScript;
+
+/**
+ * Post Install Composer Script
+ */
+class PostInstallScript extends ComposerScript {
+
+	/**
+	 * Perform the actions within this file.
+	 *
+	 * @param Event $event
+	 *
+	 * @return void
+	 */
+	public static function execute( Event $event ): void {
+		self::setEvent( $event );
+
+		// Remove the core Twenty-X themes.
+		self::deleteCoreThemes();
+	}
+
+	/**
+	 * Delete the core themes.
+	 *
+	 * @return void
+	 */
+	private static function deleteCoreThemes(): void {
+		$themes = [
+			'twentytwenty',
+			'twentytwentyone',
+			'twentytwentytwo',
+			'twentytwentythree',
+			'twentytwentyfour',
+		];
+
+		foreach ( $themes as $theme ) {
+			$theme_dir = self::translatePath( 'wp-content/themes/' . $theme );
+
+			if ( ! is_dir( $theme_dir ) ) {
+				continue;
+			}
+
+			self::deleteDirectory( $theme_dir );
+		}
+	}
+}
