+++ composer.json
@@ -39,6 +39,10 @@
     "pre-install-cmd": [
       "Viget\\ComposerScripts\\ProjectEventHandler::preInstall"
     ],
+    "post-install-cmd": [
+      "ddev start && ddev wp core download",
+      "Viget\\ComposerScripts\\ProjectEventHandler::postInstall"
+    ],
     "generate-docs": "Viget\\ComposerScripts\\WPDocsGeneratorScript::make"
   },
   "repositories": [
